Moved resolver-rs into parcel_plugin_resolver/core
