+++ crates/parcel/src/lib.rs
@@ -2,4 +2,5 @@
 
 pub mod parcel;
 pub use parcel::*;
+pub use parcel_core::*;
 pub use parcel_filesystem as file_system;
