+++ crates/parcel_core/Cargo.toml
@@ -9,7 +9,6 @@ default = []
 
 [dependencies]
 parcel_filesystem = { path = "../parcel_filesystem" }
-parcel-resolver = { path = "../../packages/utils/node-resolver-rs" }
 
 ahash = "0.8.11"
 browserslist-rs = "0.15.0"
@@ -22,3 +21,4 @@ xxhash-rust = { version = "0.8.2", features = ["xxh3"] }
 anyhow = "1.0.82"
 dyn-hash = "0.x"
 petgraph = "0.x"
+bitflags = "1.3.2"
