+++ crates/parcel_package_manager/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 [dependencies]
 mockall = "0.12.1"
 thiserror = "1.0.59"
-parcel-resolver = { path = "../../packages/utils/node-resolver-rs" }
+parcel_plugin_resolver = { path = "../parcel_plugin_resolver" }
 parcel_filesystem = { path = "../parcel_filesystem" }
+parcel_core = { path = "../parcel_core" }
 anyhow = "1.0.82"
