+++ crates/parcel_plugin_resolver/Cargo.toml
@@ -8,5 +8,22 @@ description = "Resolver Plugin for the Parcel Bundler"
 path = "./src/lib.rs"
 
 [dependencies]
+json_comments = { path = "../json-comments-rs" }
 parcel_core = { path = "../parcel_core" }
+parcel_filesystem = { path = "../parcel_filesystem" }
+
 anyhow = "1"
+typed-arena = "2.0.2"
+dashmap = "5.4.0"
+parking_lot = "0.12"
+url = "2.3.1"
+percent-encoding = "2.2.0"
+serde = { version = "1.0.152", features = ["derive"] }
+serde_json = "1.0.91"
+bitflags = "1.3.2"
+indexmap = { version = "1.9.2", features = ["serde"] }
+itertools = "0.10.5"
+elsa = "1.7.0"
+once_cell = "1.17.0"
+glob-match = "0.2.1"
+thiserror = "1.0.59"
