+++ crates/parcel_plugin_resolver/src/core/builtins.rs
