+++ crates/parcel_plugin_resolver/src/core/path.rs
