+++ crates/parcel_plugin_resolver/src/lib.rs
@@ -1,5 +1,6 @@
 #![deny(unused_crate_dependencies)]
 
+pub mod core;
 mod resolver;
 
 pub use resolver::*;
