--- packages/utils/dev-dep-resolver/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-authors = ["Devon Govett <devongovett@gmail.com>"]
-name = "parcel-dev-dep-resolver"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-parcel-resolver = { path = "../node-resolver-rs" }
-es-module-lexer = { git = "https://github.com/devongovett/es-module-lexer" }
-serde_json = "1.0.91"
-rayon = "1.7.0"
-dashmap = "5.4.0"
-glob = "0.3.1"
