--- packages/utils/node-resolver-rs/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-authors = ["Devon Govett <devongovett@gmail.com>"]
-name = "parcel-resolver"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-xxhash-rust = { version = "0.8.2", features = ["xxh3"] }
-url = "2.3.1"
-percent-encoding = "2.2.0"
-serde = { version = "1.0.152", features = ["derive"] }
-serde_json = "1.0.91"
-bitflags = "1.3.2"
-indexmap = { version = "1.9.2", features = ["serde"] }
-itertools = "0.10.5"
-json_comments = { path = "../../../crates/json-comments-rs" }
-parcel_filesystem = { path = "../../../crates/parcel_filesystem" }
-typed-arena = "2.0.2"
-elsa = "1.7.0"
-once_cell = "1.17.0"
-glob-match = "0.2.1"
-dashmap = "5.4.0"
-parking_lot = "0.12"
-thiserror = "1.0.59"
-
-[dev-dependencies]
-assert_fs = "1.0"
-
-[target.'cfg(windows)'.dev-dependencies]
-is_elevated = "0.1.2"
